fix agent and api classes relations
When classes are included or declared inside other classes, they are
not contained by default, unlike other resources. This means that if
a class A establishes a relation with a class B, A's subclasses won't
establish any relationship with B, which will be harmful when
ordering.

Without this change, to perform something before midonet_agent is
installed, one would have to refer to:

-> Class['midonet::midonet_agent::install']

which breaks encapsulation principles. After this change,

-> Class['midonet::midonet_agent'] works as expected.

Change-Id: Ie6ee75a074f7a33adb51efb60a773813b0f68827